moved handleThumbnail from News.tsx to NewsCard.tsx
